pyproject.toml
[project]
name = "explainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastapi[all]>=0.121.2",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "pydantic>=2.12.4",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.2.1",
]
